+++ src/editor/core/command/CommandAdapt.ts
@@ -2,7 +2,8 @@ import { NBSP, WRAP, ZERO } from '../../dataset/constant/Common'
 import { EDITOR_ELEMENT_STYLE_ATTR } from '../../dataset/constant/Element'
 import { titleSizeMapping } from '../../dataset/constant/Title'
 import { defaultWatermarkOption } from '../../dataset/constant/Watermark'
-import { ControlComponent, ImageDisplay } from '../../dataset/enum/Control'
+import { ImageDisplay } from '../../dataset/enum/Common'
+import { ControlComponent } from '../../dataset/enum/Control'
 import {
   EditorContext,
   EditorMode,
@@ -24,6 +25,7 @@ import {
   IGetControlValueResult,
   ISetControlExtensionOption,
   ISetControlHighlightOption,
+  ISetControlProperties,
   ISetControlValueOption
 } from '../../interface/Control'
 import {
@@ -49,8 +51,9 @@ import { IRange, RangeContext, RangeRect } from '../../interface/Range'
 import { IColgroup } from '../../interface/table/Colgroup'
 import { ITd } from '../../interface/table/Td'
 import { ITr } from '../../interface/table/Tr'
+import { ITextDecoration } from '../../interface/Text'
 import { IWatermark } from '../../interface/Watermark'
-import { deepClone, downloadFile, getUUID } from '../../utils'
+import { deepClone, downloadFile, getUUID, isObjectEqual } from '../../utils'
 import {
   createDomFromElementList,
   formatElementContext,
@@ -484,15 +487,29 @@ export class CommandAdapt {
     }
   }
 
-  public underline() {
+  public underline(textDecoration?: ITextDecoration) {
     const isDisabled =
       this.draw.isReadonly() || this.control.isDisabledControl()
     if (isDisabled) return
     const selection = this.range.getSelectionElementList()
     if (selection?.length) {
-      const noUnderlineIndex = selection.findIndex(s => !s.underline)
+      // 没有设置下划线、当前与之前有一个设置不存在、文本装饰不一致时重设下划线
+      const isSetUnderline = selection.some(
+        s =>
+          !s.underline ||
+          (!textDecoration && s.textDecoration) ||
+          (textDecoration && !s.textDecoration) ||
+          (textDecoration &&
+            s.textDecoration &&
+            !isObjectEqual(s.textDecoration, textDecoration))
+      )
       selection.forEach(el => {
-        el.underline = !!~noUnderlineIndex
+        el.underline = isSetUnderline
+        if (isSetUnderline && textDecoration) {
+          el.textDecoration = textDecoration
+        } else {
+          delete el.textDecoration
+        }
       })
       this.draw.render({
         isSetCursor: false,
@@ -1415,12 +1432,26 @@ export class CommandAdapt {
     if (!rowCol) return
     const tdList = rowCol.flat()
     // 存在则设置边框类型，否则取消设置
-    const isSetBorderType = tdList.some(td => td.borderType !== payload)
+    const isSetBorderType = tdList.some(
+      td => !td.borderTypes?.includes(payload)
+    )
     tdList.forEach(td => {
+      if (!td.borderTypes) {
+        td.borderTypes = []
+      }
+      const borderTypeIndex = td.borderTypes.findIndex(type => type === payload)
       if (isSetBorderType) {
-        td.borderType = payload
+        if (!~borderTypeIndex) {
+          td.borderTypes.push(payload)
+        }
       } else {
-        delete td.borderType
+        if (~borderTypeIndex) {
+          td.borderTypes.splice(borderTypeIndex, 1)
+        }
+      }
+      // 不存在边框设置时删除字段
+      if (!td.borderTypes.length) {
+        delete td.borderTypes
       }
     })
     const { endIndex } = this.range.getRange()
@@ -1436,12 +1467,26 @@ export class CommandAdapt {
     if (!rowCol) return
     const tdList = rowCol.flat()
     // 存在则设置单元格斜线类型，否则取消设置
-    const isSetTdSlashType = tdList.some(td => td.slashType !== payload)
+    const isSetTdSlashType = tdList.some(
+      td => !td.slashTypes?.includes(payload)
+    )
     tdList.forEach(td => {
+      if (!td.slashTypes) {
+        td.slashTypes = []
+      }
+      const slashTypeIndex = td.slashTypes.findIndex(type => type === payload)
       if (isSetTdSlashType) {
-        td.slashType = payload
+        if (!~slashTypeIndex) {
+          td.slashTypes.push(payload)
+        }
       } else {
-        delete td.slashType
+        if (~slashTypeIndex) {
+          td.slashTypes.splice(slashTypeIndex, 1)
+        }
+      }
+      // 不存在斜线设置时删除字段
+      if (!td.slashTypes.length) {
+        delete td.slashTypes
       }
     })
     const { endIndex } = this.range.getRange()
@@ -1915,6 +1960,22 @@ export class CommandAdapt {
   public changeImageDisplay(element: IElement, display: ImageDisplay) {
     if (element.imgDisplay === display) return
     element.imgDisplay = display
+    if (
+      display === ImageDisplay.FLOAT_TOP ||
+      display === ImageDisplay.FLOAT_BOTTOM
+    ) {
+      const positionList = this.position.getPositionList()
+      const { startIndex } = this.range.getRange()
+      const {
+        coordinate: { leftTop }
+      } = positionList[startIndex]
+      element.imgFloatPosition = {
+        x: leftTop[0],
+        y: leftTop[1]
+      }
+    } else {
+      delete element.imgFloatPosition
+    }
     this.draw.getPreviewer().clearResizer()
     this.draw.render({
       isSetCursor: false
@@ -1958,6 +2019,10 @@ export class CommandAdapt {
 
   public getWordCount(): Promise<number> {
     return this.workerManager.getWordCount()
+  }
+
+  public getRange(): IRange {
+    return deepClone(this.range.getRange())
   }
 
   public getRangeText(): string {
@@ -2286,6 +2351,12 @@ export class CommandAdapt {
     const isReadonly = this.draw.isReadonly()
     if (isReadonly) return
     this.draw.getControl().setExtensionByConceptId(payload)
+  }
+
+  public setControlProperties(payload: ISetControlProperties) {
+    const isReadonly = this.draw.isReadonly()
+    if (isReadonly) return
+    this.draw.getControl().setPropertiesByConceptId(payload)
   }
 
   public setControlHighlight(payload: ISetControlHighlightOption) {