+++ src/editor/core/event/handlers/mousedown.ts
@@ -1,9 +1,21 @@
+import { ImageDisplay } from '../../../dataset/enum/Common'
 import { ElementType } from '../../../dataset/enum/Element'
 import { MouseEventButton } from '../../../dataset/enum/Event'
 import { deepClone } from '../../../utils'
 import { isMod } from '../../../utils/hotkey'
 import { CheckboxControl } from '../../draw/control/checkbox/CheckboxControl'
 import { CanvasEvent } from '../CanvasEvent'
+
+export function setRangeCache(host: CanvasEvent) {
+  const draw = host.getDraw()
+  const position = draw.getPosition()
+  const rangeManager = draw.getRange()
+  // 缓存选区上下文信息
+  host.isAllowDrag = true
+  host.cacheRange = deepClone(rangeManager.getRange())
+  host.cacheElementList = draw.getElementList()
+  host.cachePositionList = position.getPositionList()
+}
 
 export function mousedown(evt: MouseEvent, host: CanvasEvent) {
   if (evt.button === MouseEventButton.RIGHT) return
@@ -20,10 +32,7 @@ export function mousedown(evt: MouseEvent, host: CanvasEvent) {
         evt.offsetY
       )
       if (isPointInRange) {
-        host.isAllowDrag = true
-        host.cacheRange = deepClone(range)
-        host.cacheElementList = draw.getElementList()
-        host.cachePositionList = position.getPositionList()
+        setRangeCache(host)
         return
       }
     }
@@ -52,7 +61,9 @@ export function mousedown(evt: MouseEvent, host: CanvasEvent) {
   // 记录选区开始位置
   host.mouseDownStartPosition = {
     ...positionResult,
-    index: isTable ? tdValueIndex! : index
+    index: isTable ? tdValueIndex! : index,
+    x: evt.offsetX,
+    y: evt.offsetY
   }
   const elementList = draw.getElementList()
   const positionList = position.getPositionList()
@@ -119,6 +130,15 @@ export function mousedown(evt: MouseEvent, host: CanvasEvent) {
     draw.getCursor().drawCursor({
       isShow: false
     })
+    // 点击图片允许拖拽调整位置
+    setRangeCache(host)
+    // 浮动元素创建镜像图片
+    if (
+      curElement.imgDisplay === ImageDisplay.FLOAT_TOP ||
+      curElement.imgDisplay === ImageDisplay.FLOAT_BOTTOM
+    ) {
+      draw.getImageParticle().createFloatImage(curElement)
+    }
   }
   // 表格工具组件
   const tableTool = draw.getTableTool()