+++ src/editor/dataset/constant/Element.ts
@@ -10,7 +10,8 @@ export const EDITOR_ELEMENT_STYLE_ATTR: Array<keyof IElement> = [
   'size',
   'italic',
   'underline',
-  'strikeout'
+  'strikeout',
+  'textDecoration'
 ]
 
 export const EDITOR_ROW_ATTR: Array<keyof IElement> = ['rowFlex', 'rowMargin']
@@ -32,7 +33,8 @@ export const EDITOR_ELEMENT_COPY_ATTR: Array<keyof IElement> = [
   'dateFormat',
   'groupIds',
   'rowFlex',
-  'rowMargin'
+  'rowMargin',
+  'textDecoration'
 ]
 
 export const EDITOR_ELEMENT_ZIP_ATTR: Array<keyof IElement> = [
@@ -64,13 +66,17 @@ export const EDITOR_ELEMENT_ZIP_ATTR: Array<keyof IElement> = [
   'listStyle',
   'listWrap',
   'groupIds',
-  'conceptId'
+  'conceptId',
+  'imgDisplay',
+  'imgFloatPosition',
+  'textDecoration'
 ]
 
 export const TABLE_TD_ZIP_ATTR: Array<keyof ITd> = [
   'verticalAlign',
   'backgroundColor',
-  'borderType'
+  'borderTypes',
+  'slashTypes'
 ]
 
 export const TABLE_CONTEXT_ATTR: Array<keyof IElement> = [
@@ -85,6 +91,12 @@ export const LIST_CONTEXT_ATTR: Array<keyof IElement> = [
   'listId',
   'listType',
   'listStyle'
+]
+
+export const CONTROL_CONTEXT_ATTR: Array<keyof IElement> = [
+  'control',
+  'controlId',
+  'controlComponent'
 ]
 
 export const EDITOR_ELEMENT_CONTEXT_ATTR: Array<keyof IElement> = [