+++ src/editor/interface/Element.ts
@@ -1,4 +1,5 @@
-import { ControlComponent, ImageDisplay } from '../dataset/enum/Control'
+import { ImageDisplay } from '../dataset/enum/Common'
+import { ControlComponent } from '../dataset/enum/Control'
 import { ElementType } from '../dataset/enum/Element'
 import { ListStyle, ListType } from '../dataset/enum/List'
 import { RowFlex } from '../dataset/enum/Row'
@@ -7,6 +8,7 @@ import { TableBorder } from '../dataset/enum/table/Table'
 import { IBlock } from './Block'
 import { ICheckbox } from './Checkbox'
 import { IControl } from './Control'
+import { ITextDecoration } from './Text'
 import { IColgroup } from './table/Colgroup'
 import { ITr } from './table/Tr'
 
@@ -30,6 +32,7 @@ export interface IElementStyle {
   rowFlex?: RowFlex
   rowMargin?: number
   letterSpacing?: number
+  textDecoration?: ITextDecoration
 }
 
 export interface IElementGroup {
@@ -100,6 +103,10 @@ export interface IDateElement {
 
 export interface IImageElement {
   imgDisplay?: ImageDisplay
+  imgFloatPosition?: {
+    x: number
+    y: number
+  }
 }
 
 export interface IBlockElement {