+++ src/editor/interface/table/Td.ts
@@ -21,8 +21,8 @@ export interface ITd {
   positionList?: IElementPosition[]
   verticalAlign?: VerticalAlign
   backgroundColor?: string
-  borderType?: TdBorder
-  slashType?: TdSlash
+  borderTypes?: TdBorder[]
+  slashTypes?: TdSlash[]
   mainHeight?: number // 内容 + 内边距高度
   realHeight?: number // 真实高度（包含跨列）
   realMinHeight?: number // 真实最小高度（包含跨列）